# APIGateway&BackendService/openapi/gateway.yaml
openapi: 3.0.3
info:
  title: Food Delivery API Gateway
  version: 1.0.0
  description: >
    Central API Gateway for the Food Delivery platform. Orchestrates routing to domain services:
    User, Hotel & Menu, Order, Payment, Review & Rating, Notification, Location, and Promotion.
    This spec mirrors the endpoints exposed by services and defines security and response contracts.
servers:
  - url: https://api.foodly.example.com
    description: Production
  - url: https://staging-api.foodly.example.com
    description: Staging
  - url: http://localhost:8080
    description: Local Gateway
...
security:
  - bearerAuth: []
paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistrationRequest'
      responses:
        '201':
          description: User registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '409':
          description: Email already exists
        '400':
          description: Invalid input
  /auth/login:
    post:
      tags: [Auth]
      summary: Authenticate user with email/password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authenticated
          headers:
            Set-Cookie:
              schema:
                type: string
              description: Optional session cookie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthToken'
        '401':
          description: Invalid credentials
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token using refresh token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: New tokens
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthToken'
        '401':
          description: Invalid refresh token
  /users/me:
    get:
      tags: [Users]
      summary: Get current authenticated user profile
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
    patch:
      tags: [Users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        '200':
          description: Updated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
  /hotels:
    get:
      tags: [Hotels]
      summary: Search and list nearby hotels
      parameters:
        - in: query
          name: q
          schema: { type: string }
          description: Search term
        - in: query
          name: lat
          schema: { type: number, format: double }
        - in: query
          name: lng
          schema: { type: number, format: double }
        - in: query
          name: radius
          schema: { type: integer, minimum: 1, default: 5000 }
          description: Meters
        - in: query
          name: cuisine
          schema: { type: string }
        - in: query
          name: ratingMin
          schema: { type: number, minimum: 0, maximum: 5 }
        - in: query
          name: sort
          schema:
            type: string
            enum: [distance, rating, popularity]
      responses:
        '200':
          description: List of hotels
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Hotel' }
  /hotels/{hotelId}:
    get:
      tags: [Hotels]
      summary: Get hotel details by ID
      parameters:
        - $ref: '#/components/parameters/HotelId'
      responses:
        '200':
          description: Hotel
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Hotel'
        '404':
          description: Not found
  /hotels/{hotelId}/menu:
    get:
      tags: [Menus]
      summary: Get menu for a hotel
      parameters:
        - $ref: '#/components/parameters/HotelId'
      responses:
        '200':
          description: Menu items
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/MenuItem' }
  /cart:
    get:
      tags: [Cart]
      summary: Get current cart
      responses:
        '200':
          description: Current cart items
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/CartItem' }
    post:
      tags: [Cart]
      summary: Replace cart with provided items
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items: { $ref: '#/components/schemas/CartItem' }
      responses:
        '200':
          description: Updated cart
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/CartItem' }
  /cart/items:
    post:
      tags: [Cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CartItem'
      responses:
        '201':
          description: Added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartItem'
  /cart/items/{itemId}:
    delete:
      tags: [Cart]
      summary: Remove item from cart
      parameters:
        - in: path
          name: itemId
          required: true
          schema: { type: string }
      responses:
        '204':
          description: Removed
  /orders:
    post:
      tags: [Orders]
      summary: Place a new order from current cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreateRequest'
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
    get:
      tags: [Orders]
      summary: List orders for current user
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [created, confirmed, preparing, picked_up, delivering, delivered, cancelled, refunded]
      responses:
        '200':
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Order' }
  /orders/{orderId}:
    get:
      tags: [Orders]
      summary: Get order by id
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: Order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Not found
    patch:
      tags: [Orders]
      summary: Update order status (authorized roles)
      parameters:
        - $ref: '#/components/parameters/OrderId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [cancelled]
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
  /payments/intent:
    post:
      tags: [Payments]
      summary: Create a payment intent for an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [orderId, method]
              properties:
                orderId:
                  type: string
                method:
                  type: string
                  enum: [card, wallet, upi, cod]
      responses:
        '201':
          description: Payment intent created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentIntent'
  /payments/{paymentId}:
    get:
      tags: [Payments]
      summary: Get payment status
      parameters:
        - in: path
          name: paymentId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Payment status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentIntent'
  /reviews:
    post:
      tags: [Reviews]
      summary: Submit a review for a hotel or order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewCreateRequest'
      responses:
        '201':
          description: Review created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
    get:
      tags: [Reviews]
      summary: List reviews for a hotel
      parameters:
        - in: query
          name: hotelId
          required: true
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, default: 20, minimum: 1, maximum: 100 }
      responses:
        '200':
          description: Reviews
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Review' }
  /promotions/validate:
    post:
      tags: [Promotions]
      summary: Validate a promo code for the current cart/order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PromotionValidationRequest'
      responses:
        '200':
          description: Promotion validation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PromotionValidationResult'
        '404':
          description: Invalid or expired promotion
  /notifications:
    post:
      tags: [Notifications]
      summary: Send a notification (admin/automation)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationCreateRequest'
      responses:
        '202':
          description: Notification accepted
...
  /location/updates:
    post:
      tags: [Location]
      summary: Submit a live location update (delivery agent)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LocationUpdate'
      responses:
        '202':
          description: Accepted
  /location/track/{orderId}:
    get:
      tags: [Location]
      summary: Get latest tracking location for an order
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: Latest location
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LocationSnapshot'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    HotelId:
      in: path
      name: hotelId
      required: true
      schema: { type: string }
    OrderId:
      in: path
      name: orderId
      required: true
      schema: { type: string }
  schemas:
...
    User:
      type: object
      required: [id, email, role, createdAt]
      properties:
        id: { type: string }
        email: { type: string, format: email }
        name: { type: string }
        phone: { type: string }
        role:
          type: string
          enum: [customer, admin, partner, courier]
        addresses:
          type: array
          items:
            $ref: '#/components/schemas/Address'
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    AuthToken:
      type: object
      required: [accessToken, refreshToken, expiresIn]
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        tokenType:
          type: string
          example: Bearer
        expiresIn:
          type: integer
          description: Seconds until access token expiry
    UserRegistrationRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
        name: { type: string }
        phone: { type: string }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    UserUpdateRequest:
      type: object
      properties:
        name: { type: string }
        phone: { type: string }
        addresses:
          type: array
          items: { $ref: '#/components/schemas/Address' }
    Hotel:
      type: object
      required: [id, name, rating, isOpen, location]
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string }
        cuisines:
          type: array
          items: { type: string }
        rating:
          type: number
          format: float
          minimum: 0
          maximum: 5
        ratingCount: { type: integer }
        priceLevel:
          type: integer
          minimum: 1
          maximum: 4
        isOpen: { type: boolean }
        location:
          $ref: '#/components/schemas/GeoPoint'
        address:
          $ref: '#/components/schemas/Address'
        etaMinutes:
          type: integer
          description: Estimated delivery time
        imageUrl: { type: string, format: uri }
    MenuItem:
      type: object
      required: [id, name, price, hotelId]
      properties:
        id: { type: string }
        hotelId: { type: string }
        name: { type: string }
        description: { type: string }
        price:
          type: number
          format: float
          minimum: 0
        currency:
          type: string
          example: USD
        isVeg: { type: boolean }
        spicyLevel:
          type: integer
          minimum: 0
          maximum: 3
        imageUrl: { type: string, format: uri }
        available: { type: boolean }
        options:
          type: array
          items:
            $ref: '#/components/schemas/MenuOptionGroup'
...
    CartItem:
      type: object
      required: [menuItemId, quantity]
      properties:
        id: { type: string }
        menuItemId: { type: string }
        hotelId: { type: string }
        quantity:
          type: integer
          minimum: 1
        notes: { type: string }
        selectedOptions:
          type: array
          items:
            type: object
            properties:
              optionId: { type: string }
        unitPrice: { type: number, format: float }
        totalPrice: { type: number, format: float }
...
    PromotionValidationRequest:
      type: object
      required: [code, orderId]
      properties:
        code: { type: string }
        orderId: { type: string }
...